dataflow: Coerce values in non-insert table ops
Because the wal reader doesn't know enough about enum types to convert
enum values to their proper ReadySet representation (the index of the
enum variant) we pass them from the replicator to tables as strings,
then coerce them to the correct representation in
`apply_table_op_coercions` in base.rs. This function, however, only
applied on *Insert* ops, and completely ignored all other
ops (including, notably, update!). This meant that if we ever *updated*
an enum value, we'd store it as a string instead of the variant index,
then fail when trying to convert it to a psql result due to it not being
an integer. This commit updates `apply_table_op_coercions` to coerce
all (row) values in all kinds of ops, covering that with a test that
updates and then queries an enum value.

Fixes: ENG-2721
Change-Id: If705dc9af703dd9ad97e24233e650b3a5e2729c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4583
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>