dataflow-expression: Fix swapped args in lowered substring call
The way this was written previously to construct an iterator of args to
pass to `BuiltinFunction::from_name_and_args` would cause a substring
call with len but not pos to get incorrectly lowered with the len arg in
the place of the pos arg - this was missed however because we don't have
tests for this expr that do both lowering and eval!

Instead, this doesn't go through `BuiltinFunction::from_name_and_args`
at all, instead opting to construct the `Substring` enum variant
directly.

Change-Id: I9ee9a0874e3cf24783663ba5c1742a3f337ed5c2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4586
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>