dataflow-expression: Make builtin eval tests go through lowering
In preparation for moving argument coercion for builtins to cast nodes
in the expression tree, refactor all tests for builtins that currently
construct an Expr directly to instead go through parsing and lowering.
This means that when we change lowering to emit coerce nodes, these
tests will continue to function properly without having to be changed to
also include coerce nodes themselves.

Change-Id: I99b5f47f80e2bfd19e6037f6223f938ef308a2ab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4587
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>