server: Mark catalog tables as non-replicated
For probably obvious reasons, we never replicate postgres catalog
tables. However, since we never even *consider* those tables, we also
don't add them to `non_replicated_relations` in the server. This means
that queries that reference those tables will get marked as permanently
"pending" in the adapter (and `SHOW PROXIED QUERIES` output), since as
far as we know the query might become supported if we eventually
replicate these tables.

For now, this commit just adds a check to see if a table is a catalog
table when deciding what error to construct for a table not being found,
and treats catalog tables the same as relations we're otherwise not
replicating.

Change-Id: Idd397a600d8b4f83d29847f09295b6e6476fff28
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4602
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI