dataflow-expression: Allow marking CASTs as fallible
Add a boolean flag to Expr::Cast which, if set to true, will make the
cast evaluate to NULL rather than returning an error if the call to
`DfValue::coerce_to` fails. This will allow us to have some of the
implicit coercions we're going to place around arguments to builtin
functions evaluate to NULL rather than failing the full evaluation if
the coercion happens to fail (for example, this is the expected behavior
for parsing the argument to the mysql `convert_tz` function).

An earlier version of this commit introduced a brand new node called
"Try" which implemented this fallibility - we're going this way instead
because we only want to handle errors caused by `coerce_to` itself, not
eg evaluating the inner expression inside the CAST.

Change-Id: I09be847d435dbf8495cee99902c0b29adb937113
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4603
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>