Extend cleanup for replication assets
This commit extends the cleanup mode to also include cleaning up
postgres replication resources and triggers. This is accomplished by the
following:
1. Drop the publication
2. Drop the replication slot
3. Drop the readyset schema

This is executed when the --cleanup command is passed. This also pipes
through a way to communicate to the caller in an async way that cleanup
has finished so we can wait appropriately before completing the rest of
the shutdown process post cleanup.

Change-Id: I84ede22b630524451cb08d793b28f14ed8ac7b75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4299
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>