domains: Handle replication offsets for uninitialized base tables.
This commit changes the way we compute and return the replication
offsets of a domain.
Previously, we returned an `Option`, and if it was `None`, that meant
for us that we needed to resnapshot the table.
Since we'll defer the initialization of the persistent state for base
table nodes in future commits, they won't have a replication offset set,
 so now we return an enum instead, which can be `Initialized`, with the
 optional offset as we had before, or `Pending`.

In practice, we are supposed to have the base tables ready before we
call `replication_offsets`, so on the client side, we throw an error if
we encounter a `Pending` for a given table.

Fixes: ENG-2352
Change-Id: I12922f8fc02c5ed7790b90732c5eb9e2a80cddfa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4611
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>