Carry column-placeholder BinaryOperators to reader
Removes the get_select_statement_binops() pass in the adapter, and
instead carries the BinaryOperators used in column-placeholder
comparisons to the Reader (and then View) by adding them to the
ViewPlaceholder::OneToOne enum. ReaderHandle::key_mapping now has all
the information on binary operators that is required to build the
ViewQuery.

ReaderHandle::build_view_query() and QueryGraph::view_key() are slightly
refactored to support the new semantics.

This change fixes a bug that prevented us from building a `ViewQuery`
for an inlined cache when the inlined statement had different types of
column-placeholder binary comparisons than the statement.

Change-Id: I5e0d0dbfb2d439e50f10730431d41c4be1dc668c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4620
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>