mir: Remove Latest node
As the comment says, this is unused, and has been for quite a while. I'm
not sure why it ever existed in the first place, but I think since it's
been years we can safely get rid of it now. In a follow-up I'll also be
removing the latest dataflow node.

Change-Id: I0f9903d6c8a240ef49c9731c148e912802a81cd5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4624
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>