clustertests: improve port allocation technique
The `DeploymentBuilder` and `DeploymentHandle` previously allocated
ports through a function that required passing the last-allocated port
in order to ensure that the same port was not assigned twice. This was
a bit error prone, since it required code authors to remember to pass
the *correct* port to this function: if the port passed to the function
was not the most recently-allocated port **and** the port that was
*actually* previously allocated was not in use yet, the function could
potentially return a port that was already allocated.

To make port allocation more robust and less error prone, this change
introduces a new `PortAllocator` type that keeps *internal* state to
keep track of the ports that have already been allocated. This approach
doesn't require callers to keep track of the ports that have already
been used, preventing the possibility of ports being allocated more than
once.

Change-Id: Ied3c38745a620da3aa2fae237a0864cc1f0dd6b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4634
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>