ddl_vertical: Add timeout for individual test cases
When I found ENG-2763, it didn't actually cause a test failure, it
caused the test to hang. I came up with this timeout code to allow
shrinking to proceed in cases like this, but forgot to commit it until
now.

I wanted to use the builtin `timeout` field of `ProptestConfig` but it
turns out that option is not compatible with the closure-like syntax
that we use for our test runs, which in turn is needed for us to be able
to pass in `next_step_idx`. Using `tokio::time::timeout` with `unwrap()`
causes us to panic if the timeout is exceeded, which fails the test case
and produces the same end result.

Change-Id: Ie9a4222127e63ebfcfc07776439e0355f38fe5e5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4631
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI