readyset: Add --no-upstream-connections flag
Add a flag which if set will disable making connections upstream per
client connection (even if we *do* have an upstream db url configured
for replicators). This prevents proxying queries if they fail (which we
already support due to being able to just set the upstream to None) but
allows us to accept considerably more connections due to not hitting
PostgreSQL's client connection limit.

This flag is primarily for use in standalone mode - obviously when
running a separate server and adapter we can just not specify the
upstream db url as a flag.

Release-Notes-Core: Added a new flag, `--no-upstream-connections`, which
  will prevent ReadySet from making connections to the upstream db for
  every client connection. This disables proxying unsupported queries,
  but allows ReadySet to accept considerably more client connections due
  to not being limited by the upstream database's connection limit.
Change-Id: I69506fedc6fc06c74af215211434f27e4b9b386f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4633
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>