replicators: Skip catch up if ReadySet did not snapshot any table.
Currently, every time we start ReadySet, we request the current
replication offset (the earliest location from which streaming can
start) from the Postgres replication log, and we _always_ set the schema
 replication offset to that value.
In doing so, we almost certainly always end up triggering a catch-up
process, which will essentially try to go from the lowest replication
log we have registered, to the replication offset returned by Postgres.

The problem that if no change was made to the schema (tables and
views remained the same) from the last time ReadySet was run, then
there's no need to catch up, as nothing really changed.

This commit allow us to skip the catch-up phase entirely, if:
1. ReadySet has already been deployed in the past (by checking if the
have a replication offset already assigned to the schema), and
2. No snapshot was triggered for any table.

To skip the catch-up phase, we simply avoid overriding the current
schema replication offset with the offset provided by Postgres.

Fixes: ENG-2785
Change-Id: I39bd5b0c08c87cc69af5a03821f02754b337866a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4637
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>