dataflow-expression: Decide coercion for array_to_string at lowering
Just like we're doing for all the other functions, move the coercion of
the arguments to the `array_to_string` function from happening within
the evaluation of the builtin to happening as explicit CAST expr nodes
created at lowering time.

This doesn't include new tests in postgres_oracle because postgres is
actually *more* strict than we have to be and requires the arguments are
already known to be an array - this obviously means that we wouldn't be
testing casting behavior in these tests, so I've decided that the
existing unit tests are good enough.

Change-Id: I5cabdc62578e92320c60d84b4ec18586d1aa6e95
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4657
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>