dataflow-expression: Don't cast if we don't need to
As an optimization, if we're trying to cast an expr that's an argument
to a builtin function, and the type we're casting to is already the type
of the expression, just omit the cast node entirely, putting the
internal expr in its place instead. I haven't actually measured the
impact of this, but it's so cheap of an optimization that we might as
well do it just to save on the extra expression tree depth.

Change-Id: I5fb73a514869a74da54840fa1269ba2b64b5014e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4659
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>