treewide: Upgrade to clap 4.2.1
Upgrade the clap crate everywhere to version 4.2.1, updating usage of
backwards-incompatible APIs along the way. Some highlights:

- Many enums are now just deriving ValueEnum, as the new version of the
  previous method of using `FromStr` and `possible_values` is much
  clunkier than before.
- Some values used in clap args now are required to derive Clone
- We have to either avoid automatically creating arg groups for structs
  that are `#[clap(flatten)]`ed, or give them a unique name with
  `#[group(id = ...)]`, or clap will give us an error that the groups
  are not unique.
- Arg names now use underscores instead of dashes when referenced from
  other args

Also, to keep the single-version-policy intact, a couple of other
dependencies had to update:
- rustyline in readyset_repl, to update linux-raw-sys
- console treewide
- indicatif treewide

Unfortunately the latest version of nperf-core depends on an older
version of the `syn` crate than clap uses, so there's nothing we can do
there and we just have to ignore it here.

Change-Id: I4f353294c85ebd1d494537baa16d3b2e797bec69
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4660
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>