adapter: Only redact literals in SHOW CAHCES if feature is enabled
Make the redaction of literals in the output of SHOW CACHES conditional
on the `redact_sensitive` feature being enabled, which is the case for
all other places we redact literals. In deployments that don't care
about PII, this is nice since it allows directly copy-pasting queries
from `SHOW CACHES` to run them (modulo parameters)

As a side note, we should probably work to make this whole thing a
runtime flag rather than a compile-time feature flag so we don't have to
build two binaries, but that's a separate piece of work and one we
already need to do.

Change-Id: I7198f60102396d62aa7386596ed02d9aefb4c391
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4664
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>