sql-passes: Don't anonymize NULL or booleans
The point of the literal anonymization pass is to scrub queries that
might contain inlined literals with sensitive data such as PII in them
to allow them to be logged verbatim or sent back to ReadySet over
telemetry. This makes sense for string literals and even integer
literals, but (at least in my opinion) it's unlikely that any
information could be leaked via a boolean or a NULL literal in a query.
This commit changes the anonymize_literals pass to skip anonymizing
those literals.

Change-Id: Ie1615a3bc790967392ac0e55722439cd7fa8faca
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4665
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>