mir: Add a rewrite pass to add a bogokey to a query
This commit starts the preparation for having the parameters used in a
query be determined more dynamically and mutably via rewrite passes on
MIR, rather than entirely statically at QG->MIR lowering, but adding a
new (unused!) rewrite pass that determines if a query needs a
bogokey (because it doesn't have any keys of its own) and if so adds a
bogokey project node above the leaf in addition to a bogokey in the
leaf. The idea here is that because we might be changing the parameters
in a query dynamically via other rewrite passes (eg to pull parameters
up out of subqueries into the leaf), we might not know if we need a
bogokey until the end of rewrite, so it's too early to make that
decision at lowering time.

Change-Id: I72234471e48fcb1b3586731525742447a9e909f2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4668
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>