replicators: batch last few snapshot RPCs
Previously, after inserting the last batch of rows from the snapshot
into the base table, we'd send two more RPCs: one to set the replication
offset on the table and one to disable snapshot mode (and, thus, compact
the table). In the event that we successfully set the replication offset
but fail to disable snapshot mode (e.g. due to transient network
issues), we would not be able to successfully recover upon retrying:
once the replication offset on the table is set, we assume the table has
been successfully snapshotted and never revisit it to disable snapshot
mode. The table ends up stuck in the `Snapshotting` state and is never
made queryable.

This change batches the RPCs to set the replication offset and disable
snapshot mode along with the RPC that sends the last batch of rows. This
ensures that the table never ends up in an inconsistent state where it
has every row and a replication offset but is stuck in snapshot mode.

Change-Id: I2991cd79bb3b5e66631bc8a4105374c72a41ff21
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4670
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>