treewide: Boolean flags now need to be "true", not "1"
With the upgrade to clap 4.2.1 in d9c4e8994 (treewide: Upgrade to clap
4.2.1, 2023-04-07), in order to enable boolean flags from enviroinment
variables we now need to set that environment variable to the string
"true" instead of "1". This updates all configuration treewide that I
could find that was enabling boolean flags, most notably the
docker-compose files used by the framework tests in the nightly build,
which was breaking after the Clap upgrade.

Change-Id: I2f4d321b49af91038f4a55678ac476bf930487e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4672
Reviewed-by: Ethan Donowitz <ethan@readyset.io>
Tested-by: Buildkite CI