mir: Add a new ViewKey node type
Add a new `ViewKey` node type to MIR, which will be used to represent
filter parameters in queries that aren't *trivially* convertible to keys
in the Leaf, eg in subqueries. The idea is that we'll create these nodes
during compilation to MIR, then pull them down to the Leaf node (if
possible) in a MIR rewrite pass - after that, any that remain will cause
the whole query to return an `UnsupportedPlaceholders` error during
lowering to dataflow.

Currently, however, this is all hypothetical, and this node type is
completely unused.

Change-Id: I691ea4b7bc3b4560a0d044cadd8b1c8ed4b2085c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4674
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>