nom-sql: Don't blow the stack when generating SqlType
SqlType::Array makes the SqlType AST enum in nom-sql a recursive type -
by default, the derived impl for Proptest::Arbitrary will attempt to
generate arbitrarily nested arrays, which basically means that any
attempt to generate a SqlType will just cause a stack overflow. We don't
actually want to generate arrays-of-arrays here (that's not how nested
arrays work in postgresql). Unfortunately, the `test_strategy` derive
macro for `proptest::Arbitrary` doesn't support skipping enum variants
entirely based on parameters supplied to the generator, so to skip
generating arrays inside of arrays we have to convert this to a manual
implementation.

Change-Id: Ia5f4afa4f060f21a37fb1559c876467adec7837a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4677
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>