ddl_vertical: Add InsertEnumValue command
This is similar to the AppendEnumValue command, in that it also alters
an enum type and adds a value. However, instead of defaulting to adding
the new value to the end of the enum, it adds a value to an arbitrary
point in the list of existing values. This can mean either specifying
"BEFORE" or "AFTER" a value in the "ALTER TYPE" statement.

Refs: ENG-2774
Change-Id: I24387abe78bb331c17733362de6a90cfd250bc2d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4692
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>