sql-passes: Handle subqueries in count(*) rewrite
Previously, the count(*) rewrite pass didn't know about subquery
schemas, meaning that it'd fail due to not being able to find the schema
for the subquery it selected for its column if a query *only* selected
FROM subqueries. This commit fixes that, by tracking subquery schemas
during visit in the same way we do for the expand-implied-tables pass.

Change-Id: I18a34b222099a3588d08b56ce265b32877e4ba22
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4696
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>