mir: Preserve edge weights in swap_with_child
Previously when re-inserting edges as part of MirGraph::swap_with_child,
we'd always insert an edge weight of 0. This meant that if we ever
swapped a node with its *join* child, we'd lose the edge ordering on
that join's parents, making it randomly fail to lower to dataflow due to
not being able to resolve conflicts.

This commit updates swap_with_child to preserve the edge weights of
edges, specifically, preserving the edge weights such that this use case
works correctly.

Change-Id: I4bcfd8e60a8ef6b79bd9bc5c6f91e6f4e3f2e5af
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4697
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>