sql: Make ViewKey nodes for placeholders in subqueries
When compiling a subquery to MIR, make ViewKey nodes for all the
parameters of the query, rather than just skipping them entirely. These
still return an unsupported placeholders error when lowering to
dataflow, but later on can be pulled down to be added to the leaf node
as part of a future rewrite pass.

Change-Id: I2ff7704dd4ed10dcc14f4cf1b8d982d37816061d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4698
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>