adapter: Look at subqueries for parameter binops
Dig not just into the top-level where clause of the query, but all where
clauses of all queries, when figuring out parameter binops in the
adapter. This bit generally doesn't need to be nearly as careful, since
if there is actually an unsupported placeholder it'll fail during
compilation in the server - and having this here paves the way for
actually *supporting* placeholders in subqueries.

Change-Id: Icf224d91c17a91b3b461317ca4c5e5ee942613eb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4700
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>