mir: Add a pass for pulling ViewKey nodes down to the leaf
Add a new pass to MIR called `pull_keys`, which attempts to pull all
ViewKey nodes down to the leaf node of the query to be added to the
overall key for the query, updating things like aggregates to add the
LHS of the key as a group_by along the way.

The result of this is that at least for equal placeholders (for now), we
now support parameters in (most) subqueries!

Change-Id: Ie8da5db2b43ea6de1fb131be5002de1e33c5d2ea
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4701
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>