Replicator: flush buffered actions on error
Ensures that we flush buffered actions when we receive an error in
PostgresWalConnector::next_action(). Previously, we returned the error
and dropped anything in `actions`, unless the error was an
`UnsupportedTypeConversion`.

This behavior is implemented by storing either an event or a WalError in
`PostgresWalConnector::peek`. The behavior is tested by a new
readyset-psql fallback test.

Co-authored-by: Ethan Donowitz <ethan@readyset.io>
Fixes: ENG-2832
Change-Id: I0de405d1239bd66458b53cd7648276c62baa9a74
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4708
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>