adapter: Shut down the embedded server task
When ReadySet runs in standalone mode, the ReadySet server is embedded
in the same process as the adapter as a separate task. Previously, we
had not been shutting down the embedded server task when the adapter shut
down. This commit ensures that, when the adapter shuts down, we are calling
`ShutdownSender::shutdown` on the `ShutdownSender` that was returned
with the server handle to gracefully shut down the server's background
tasks.

Change-Id: I4aa2346ead9a2380c7ddd68563c4562aff252e3f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4710
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>