replicators: Properly handle creating tables with FKs
When an event trigger is executed for a `CREATE TABLE` statement that
also creates other entries in pg_class, such as constraints for foreign
keys, the `pg_event_trigger_ddl_commands()` function returns rows
for *all* those entries, rather than just the initial table that was
created. If the table didn't have a primary key and therefore we decided
to try to set its `REPLICA IDENTITY`, previously we'd fail due to a
subquery in that clause returning a row for each of those rows in
pg_class, rather than just the table that was created. This commit fixes
that, also updating the way the `REPLICA IDENTITY` stuff works to be
more robust to `pg_event_trigger_ddl_commands()` returning more than one
row in general.

Fixes: ENG-2829
Change-Id: Ic126d155e164164d748064c990f33f643043a7e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4723
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>