stateful-proptest: Add new crate for stateful property testing
This is based off of the code developed for the vertical DDL tests, and
is meant to offer a generic framework for building more tests like that
in the future.

For now, much of the code in the DDL vertical test suite is duplicated
here, but in subsequent commits I will port the DDL vertical tests to
use this crate.

Refs: ENG-2838
Change-Id: I08c1d79267d91cced4720aeef66e84a4a0bd63b7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4768
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>