query-generator: Don't generate bools for unique columns
If a parameter comparing to a column is used as a key, we try to
generate unique values of the type of that column to use as the lookup
key for a query. This works most of the time, but booleans in particular
aren't able to be unique enough to allow us to generate unique values
for them - meaning if we end up with a boolean column as our parameter
column, we'll panic when trying to generate the key for the query.

This commit allows ops to request for columns in a table that pass some
filter, rather than just arbitrary columns, and then uses that to make
sure that columns that we're using for parameter comparisons are never
booleans (or arrays or Other, while we're at it, since the same
limitation applies there too)

Change-Id: Iabac1e67acd4de100bc5be88da785b18f935c59c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4727
Reviewed-by: Nick Marino <nick@readyset.io>
Tested-by: Buildkite CI