logictest: Add a --verbose flag to verify
Add a verbose flag to the readyset-logictest verify subcommand, which
causes the logictest script runner to log all statements and queries
that're being executed. Then, set this flag to `true` from the generate
subcommand if the generate subcommand's `verbose` flag is itself set.
This makes debugging panics in generated logictests easier, by logging
each statement we run as we're running it.

Change-Id: Ia6e099f30419190c0e18e8491b8a2ccf2a30dc59
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4731
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>