query-generator: Ensure generated values are always valid
Ensure that `value_of_type`, `unique_value_of_type`, and
`random_value_of_type` in QueryGenerator always generate values that can
successfully be inserted into the database (starting with MySQL for now,
but PostgreSQL will come soon). This is all covered with a proptest for
each, which generates a value using these functions and attempts to
insert it into a table. For the case of `random_value_of_type`, this
required making it take the random generator as an argument rather than
initializing it directly, so we can pass in a deterministic mock random
generator in tests.

Change-Id: I86b256f589c75408b90ad8092ac95f855c31897e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4734
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>