nom-sql: Allow constraining which dialect to generate types for
Add an optional `Dialect` field to the Arbitrary impl for SqlType which
allows constraining which dialect to generate types for. Since currently
PostgreSQL is a strict superset of MySQL for our purposes, this
currently just opts-out of generating a bunch of PostgreSQL-specific
types if the dialect is not set to Some(PostgreSQL) - but I've left it
as Option<Dialect> in case in the future we want to support some
mysql-only types too.

Change-Id: I150a869ec817ae48d2a94712f1fd20875d554feb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4736
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>