logictest: Discard fuzz seeds which fail on upstream
Rather than panicking the whole test run if a fuzz seed fails to run on
the upstream database (which despite a lot of effort can still
hypothetically happen, since generating 100%-valid queries and data is a
very hard problem), just discard them using a `prop_filter_map`. This is
currently happening rarely enough that we're not throwing away a *ton*
of seeds, and we're still able to eventually generate valid test
scripts.

Change-Id: I003eb1129fb30e952e5ae1891160606da12c86a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4738
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>