ddl_vertical: Remove print_cases test
I haven't used this in ages; it was mostly helpful during the early
stages of developing the shrinking framework and whatnot. Once we
convert this test suite to use the new stateful-proptest crate, this
test case will no longer be easy to express, so I'm getting it out of
the way now so as to make the following commits smaller and simpler.

Refs: ENG-2838
Change-Id: I8ef60589fb2c6df83181d25888cb84e779fd69fc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4770
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>