adapter: Report why something is unsupported without upstream
Previously, the error message for something failing (either due to
failing to rewrite, or due to an unsupported error) when running without
an upstream (as we do when eg running logictests) wouldn't include
the *reason* why that failed - it'd just say "operation not implemented
yet". This commit tracks the actual error that occurred as part of the
PrepareMeta, and reports it as the error message that we return to the
client.

Change-Id: I8ac1c479348fc2988d3f9c44ba01772348b4b755
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4753
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>