ddl_vertical: Convert run_cases test to use stateful-proptest
This change means that the DDL vertical test no longer actually tests
anything, since more work is needed to convert the rest of the code to
use the new stateful-proptest framework. However, it does build and run,
so I'm starting with this change alone, so as to make the conversion
process easier to follow in review. Subsequent CLs will convert the rest
of this test suite.

This also removes a bunch of code that now lives in stateful-proptest,
and makes DDLModelState implement the ModelState trait.

The code changes around `preconditions` is a little bit tricky because
we change it from a method on Operation to one on DDLModelState, but the
logic is fundamentally the same.

To make the diff more readable, this commit deliberately avoids moving
functions around within the file, since doing so would make it hard to
see what code changes were made within a given function. Instead we use
stubs in the impl block, and I will do all the code relocation in the
following commit.

Refs: ENG-2838
Change-Id: I380caa6322a161b9d9a55f407baabc8f25cdd534
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4771
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>