debug: Extend HTTP errors to include the request.
We've seen a few times already that ReadySet fails due to timeouts.
For RPC timeouts, the information we have is usually enough since we
have several endpoints that handle one specific request each.
However, for HTTP requests (specifically those that target the
`/worker_request` endpoint), we have several different requests we might
 send, and we don't get any information about them on the error messages
 when these requests fail.

This commit adds a `request` string field to the
`ReadySetError::HttpRequestFailed` + adds the (debug) string
representation of the request kind to the error.

Change-Id: I0a13e41515a4b9cc487cf45c1d84b2ef87968a0b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4757
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>