logictest: Still call shutdown() if test fails
Previously, we'd short-circuit due to `?` if running a test script
failed, which prevented us from calling `shutdown_tx.shutdown().await` -
we still want to make sure that happens if we're about to fail due to an
error, though. This is especially important when running fuzz, since
that can call this function hundreds to thousands of times.

Change-Id: Ia2894a0f624a0fc2bbaa86d895df8f5fd99f70aa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4762
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>