psql-srv: Fix error message string interpolation
We had mistakenly left out curly braces around the word "username" in
the error message we print out when authentication fails for a
particular username, which caused us to display the literal "username"
in the error message instead of the actual username for which
authentication failed. This commit adds the curly braces to fix the
string interpolation.

Change-Id: I0970057966f40bf6267677ea019e05f04dbd5730
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4778
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI