readyset-repl: Allow providing types to prepare
In Postgres, types can be explicitly provided for a subset of the
parameters in a prepared statement. To test this behavior against both
ReadySet and PostgreSQL itself, this commit adds support for providing
type OIDs to the `prepare` command in readyset-repl, which delegate to
the `prepare_typed` API in tokio_postgres.

Change-Id: I4965034e61138a6abcec8795d6e96da93a7a6440
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4788
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>