readyset-psql: Pass client-provided param types to upstream
PostgreSQL allows explicit types to be provided for the placeholders in
a prepared statement. That capability is not widely used by applications
and client libraries, but it appears to be the case that at least JDBC
will use this capability, and then rely on its correct implementation
in the server by sending values with the types *it* specified, ignoring
the types the server responded with. Previously, however, we'd just
completely ignore the types provided by the client as part of the Parse
message, relying entirely on the types inferred for the statement by the
upstream database or the ReadySet server.

This commit starts us partially down the road of actually using these
types if provided, by passing them all the way through to the prepare
request we send to the upstream database for proxied queries. To support
this, there's a new `PrepareData` associated type on `UpstreamDatabase`,
which is `()` for MySQL since MySQL doesn't support this behavior, and
`&[Type]` for postgres. This is passed all the way through from the
original Parse message to the `prepare` method on `PostgresUpstream`,
which now uses `prepare_typed` instead of `prepare`. The only place
this *isn't* passed is in the upstream querying we're doing in the
MigrationHandler - there, we're just using the Default value, since
threading through extra types there is annoying and we're going to be
getting rid of that code anyway.

This isn't all the work that's necessary to get this working - we still
need to also provide these types to ReadySet itself when preparing
non-proxied queries, which I suspect will be a little bit more involved.

Change-Id: Ia257ecbf1434bcd86fb9815b7e6555471bc7571a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4789
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>