replicators: Mark partitioned tables as non-replicated
We don't currently support PostgreSQL partitioned tables - and the *way*
we don't support them is that we will replicate both the table (at least
if it's created post snapshotting) and its partitions, but not actually
do the proper query rewrite to allow querying from the table to hit the
partitions - effectively, this means we'll always return empty result
sets for queries that hit partitioned tables. Instead, this commit
always marks partitioned tables as non-replicated, both in
snapshotting (where we were already not snapshotting them because they
have relkind `p` instead of `t`) and in DDL replication. Note that
we *are* still replicating the partitions themselves, since they're just
regular tables.

Fixes: ENG-2857
Change-Id: Ie6d6cb25b03365b17d3c25d14e5df7b3a206cd76
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4800
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Ethan Donowitz <ethan@readyset.io>