sql: Put DISTINCT before leaf project node
If a (non-injective) expression returns the same value given two
different values as input, we want only one row rather than two -
previously, since the DISTINCT node was placed *before* the leaf project
node, we'd do a DISTINCT on the input columns but not the actual result
values of the leaf projected expressions. Now, we always put DISTINCT
last, which fixes the issue.

Fixes: ENG-2959
Change-Id: I2b13f3ac692fb56f112d4726df2b6c72688f6ec8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4801
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>